reorganise benchmarks
