+++ runner/config.js
@@ -5,29 +5,20 @@ const os = require('os')
 const path = require('path')
 const YAML = require('yaml')
 const Influx = require('influx')
+const util = require('util')
 const Pino = require('pino')
 const PinoPretty = require('pino-pretty')
 const pinoms = require('pino-multi-stream').multistream
 const PinoGetPrettyStream = require('pino/lib/tools').getPrettyStream
-const uuidv1 = require('uuid/v1')
-const util = require('util')
 const mkDir = util.promisify(fs.mkdir)
 
 let pino
 
+const configBenchmarks = require('./lib/configBenchmarks')
 const inventoryPath = process.env.INVENTORY || path.join(__dirname, '../infrastructure/inventory/inventory.yaml')
 const playbookPath = path.join(__dirname, '../infrastructure/playbooks/benchmarks.yaml')
-const remoteTestsPath = process.env.REMOTE_FOLDER || '~/ipfs/tests/'
-const remoteIpfsPath = process.env.REMOTE_FOLDER || '~/ipfs/'
-const tmpOut = '/tmp/out'
-const params = `OUT_FOLDER=${tmpOut} REMOTE=true GUID=${uuidv1()} `
-const remotePreNode = `killall node 2>/dev/null; killall ipfs 2>/dev/null; source ~/.nvm/nvm.sh && `
 const HOME = process.env.HOME || process.env.USERPROFILE
 const keyfile = path.join(HOME, '.ssh', 'id_rsa')
-const tests = []
-const locations = ['local', 'remote']
-const clinicOperations = ['doctor', 'flame', 'bubbleProf']
-const clinicFilesets = ['One4MBFile', 'One64MBFile']
 const memorySuffix = '_memory'
 
 const ipfsAddress = process.env.IPFS_ADDRESS || '/dnsaddr/cluster.ipfs.io'
@@ -75,165 +66,11 @@ const getBenchmarkHostname = () => {
   return getInventory().all.children.minions.hosts
 }
 
-const getCommand = (test, loc) => {
-  return `${loc === 'remote' ? remotePreNode : ''} ${testDefaults.params} node ${testDefaults.path[loc]}/${test.file}`
+let loc = 'remote'
+if (process.env.STAGE === 'local') {
+  loc = 'local'
 }
-
-const getClinicCommands = (test, operation, loc) => {
-  if (locations.includes(loc) && clinicOperations.includes(operation)) {
-    let variations = []
-    for (let fileSet of clinicRuns[operation].fileSets) {
-      let shellCommand = `${loc === 'remote' ? remotePreNode : ''} FILESET="${fileSet}" clinic ${operation} --dest ${tmpOut}/${test.name}/ -- node ${testDefaults.path[loc]}/${test.file}`
-      variations.push({
-        command: shellCommand,
-        fileSet: fileSet,
-        benchmarkName: test.name,
-        operation: operation
-      })
-    }
-    return variations
-  } else {
-    throw Error(`getClinicCommands requires an operation from ${clinicOperations} and a location from ${locations}`)
-  }
-}
-
-const clinicRuns = {
-  doctor: {
-    fileSets: ['One4MBFile', 'One64MBFile']
-  },
-  flame: {
-    fileSets: ['One4MBFile', 'One64MBFile']
-  },
-  bubbleProf: {
-    fileSets: ['One4MBFile']
-  }
-}
-
-const testDefaults = {
-  path: {
-    remote: remoteTestsPath,
-    local: path.join(__dirname, '/../tests')
-  },
-  params: params,
-  remotePreCommand: remotePreNode
-}
-
-const testAbstracts = [
-  {
-    name: 'localTransfer_tcp_mplex',
-    file: 'local-transfer.js -t tcp -m mplex'
-  },
-  {
-    name: 'localTransfer_ws_mplex',
-    file: 'local-transfer.js -t ws -m mplex'
-  },
-  {
-    name: 'localTransfer_ws_mplex',
-    file: 'local-transfer.js -t ws -m mplex'
-  },
-  {
-    name: 'localTransfer_tcp_mplex_secio',
-    file: 'local-transfer.js -t tcp -m mplex -e secio'
-  },
-  {
-    name: 'localTransfer_ws_mplex_secio',
-    file: 'local-transfer.js -t ws -m mplex -e secio'
-  },
-  {
-    name: 'localTransfer_tcp_spdy',
-    file: 'local-transfer.js -t tcp -m spdy'
-  },
-  {
-    name: 'localTransfer_ws_spdy',
-    file: 'local-transfer.js -t ws -m spdy'
-  },
-  {
-    name: 'localTransfer_tcp_spdy_secio',
-    file: 'local-transfer.js -t tcp -m spdy -e secio'
-  },
-  {
-    name: 'localTransfer_ws_spdy_secio',
-    file: 'local-transfer.js -t ws -m spdy -e secio'
-  },
-  {
-    name: 'unixFsAdd',
-    file: 'local-add.js'
-  },
-  {
-    name: 'unixFsAddTrickle',
-    file: 'local-add.js trickle'
-  },
-  {
-    name: 'localExtract',
-    file: 'local-extract.js'
-  },
-  {
-    name: 'multiPeerTransfer',
-    file: 'multi-peer-transfer.js'
-  },
-  {
-    name: 'addMultiKb',
-    file: 'add-multi-kb.js'
-  },
-  {
-    name: 'addMultiKbTrickle',
-    file: 'add-multi-kb.js trickle'
-  },
-  {
-    name: 'initializeNodeBrowser',
-    file: 'init-node.browser.js'
-  },
-  {
-    name: 'unixFsAddBrowser',
-    file: 'local-add.browser.js'
-  },
-  {
-    name: 'addMultiKbBrowser',
-    file: 'add-multi-kb.browser.js'
-  },
-  {
-    name: 'unixFsAddGo',
-    file: 'local-add.go.js'
-  },
-  {
-    name: 'extractJs2Go',
-    file: 'extract-js2.go.js'
-  },
-  {
-    name: 'extractGo2JsWs',
-    file: 'extract-go2.js -t ws'
-  },
-  {
-    name: 'extractJs2GoWs',
-    file: 'extract-js2.go.js -t ws'
-  },
-  {
-    name: 'extractGo2Js',
-    file: 'extract-go2.js'
-  },
-  {
-    name: 'peerTransferBrowser',
-    file: 'peer-transfer.browser.js'
-  },
-  {
-    name: 'pubsubMessage',
-    file: 'pubsub-message.js'
-  }
-]
-
-for (let test of testAbstracts) {
-  let loc = 'remote'
-  if (process.env.STAGE === 'local') {
-    loc = 'local'
-  }
-  tests.push({
-    name: test.name,
-    benchmark: getCommand(test, loc),
-    doctor: getClinicCommands(test, 'doctor', loc),
-    flame: getClinicCommands(test, 'flame', loc),
-    bubbleProf: getClinicCommands(test, 'bubbleProf', loc)
-  })
-}
+const tests = configBenchmarks.constructTests(loc)
 
 const runClinic = (process.env.CLINIC && (process.env.CLINIC === 'ON' || process.env.CLINIC === true)) || false
 
@@ -243,7 +80,7 @@ const config = {
   },
   log: pino,
   stage: process.env.STAGE || 'local',
-  outFolder: process.env.OUT_FOLDER || tmpOut,
+  outFolder: process.env.OUT_FOLDER || configBenchmarks.tmpOut,
   dataDir: process.env.DATADIR || './data/',
   logFile: logFile, // where we store all the stuff that is to be sent to IPFS
   now: now,
@@ -253,11 +90,11 @@ const config = {
     apikey: process.env.API_KEY || 'supersecret',
     api: {
       clinic: {
-        operations: clinicOperations,
-        filesets: clinicFilesets
+        operations: configBenchmarks.clinicOperations,
+        filesets: configBenchmarks.clinicFilesets
       },
       benchmarks: {
-        tests: testAbstracts
+        tests: configBenchmarks.testAbstracts
       }
     }
   },
@@ -311,15 +148,15 @@ const config = {
     user: process.env.BENCHMARK_USER || 'elexy',
     key: process.env.BENCHMARK_KEY || keyfile,
     path: path.join(__dirname, '../tests'),
-    remotePath: remoteTestsPath,
+    remotePath: configBenchmarks.remoteTestsPath,
     tests: tests,
-    cleanup: `rm -Rf ${tmpOut}/*`,
+    cleanup: `rm -Rf ${configBenchmarks.tmpOut}/*`,
     measurements: {
       memory: memorySuffix
     }
   },
   ipfs: {
-    path: remoteIpfsPath,
+    path: configBenchmarks.remoteIpfsPath,
     network: {
       address: ipfsAddress,
       user: ipfsUser,
